+++ src/rng/xoshiro/xoshiro512.rs
@@ -1,5 +1,4 @@
-use crate::rng::R64;
-
+use crate::rng::core::R64;
 struct Xoshiro512 {
     state: [u64; 8],
 }
